# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ascenda API",
    "version": "1.0.0",
    "description": "Hotel search & booking backend"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/api/bookings": {
      "get": {
        "summary": "List current user bookings",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "payment_id": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "payment_id"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a booking (initiates Stripe payment intent)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination_id": {
                    "type": "string",
                    "example": "EzoR",
                    "description": "Ascenda destination UID"
                  },
                  "hotel_id": {
                    "type": "string",
                    "example": "MU74",
                    "description": "Ascenda hotel UID"
                  },
                  "start_date": {
                    "type": "string",
                    "example": "2026-08-24",
                    "description": "ISO date string"
                  },
                  "end_date": {
                    "type": "string",
                    "example": "2026-08-26",
                    "description": "ISO date string"
                  },
                  "adults": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991,
                    "example": 2,
                    "description": "Number of adults"
                  },
                  "children": {
                    "default": 0,
                    "example": 0,
                    "description": "Number of children",
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "message_to_hotel": {
                    "example": "Late check-in preferred",
                    "type": "string"
                  },
                  "room_types": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "e9e8555f-e187-5194-819a-45bfa9ca9cec"
                    ]
                  },
                  "price_paid": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "example": 901.04,
                    "description": "Total amount in SGD"
                  },
                  "payee_id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "payment_id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "user_id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "description": "Supabase user UUID"
                  }
                },
                "required": [
                  "destination_id",
                  "hotel_id",
                  "start_date",
                  "end_date",
                  "adults",
                  "room_types",
                  "price_paid",
                  "user_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns client_secret for Stripe confirmation"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/hotels": {
      "get": {
        "summary": "Search hotels by destination",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "destination_id",
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9]{4,64}$",
              "example": "EzoR",
              "description": "Ascenda destination UID"
            },
            "required": true,
            "description": "Ascenda destination UID"
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel listing"
          },
          "400": {
            "description": "Missing or invalid destination_id"
          },
          "502": {
            "description": "Upstream Ascenda API unavailable"
          }
        }
      }
    },
    "/api/hotels/prices": {
      "get": {
        "summary": "Fetch priced hotel availability",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "destination_id",
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "EzoR"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "checkin",
            "schema": {
              "type": "string",
              "pattern": "^(0[1-9]|1[0-9]|20)[0-9]{2}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$",
              "example": "2026-08-24",
              "description": "ISO date string"
            },
            "required": true,
            "description": "ISO date string"
          },
          {
            "in": "query",
            "name": "checkout",
            "schema": {
              "type": "string",
              "pattern": "^(0[1-9]|1[0-9]|20)[0-9]{2}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$",
              "example": "2026-08-26",
              "description": "ISO date string"
            },
            "required": true,
            "description": "ISO date string"
          },
          {
            "in": "query",
            "name": "guests",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "2",
              "description": "Total guest count"
            },
            "required": true,
            "description": "Total guest count"
          },
          {
            "in": "query",
            "name": "currency",
            "schema": {
              "default": "SGD",
              "example": "SGD",
              "type": "string",
              "pattern": "^[A-Z]{3}$"
            }
          },
          {
            "in": "query",
            "name": "country_code",
            "schema": {
              "default": "SG",
              "example": "SG",
              "type": "string",
              "pattern": "^[A-Z]{2}$"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "default": "en_US",
              "example": "en_US",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pricing data with completed flag"
          },
          "400": {
            "description": "Invalid parameters or dates"
          },
          "502": {
            "description": "Upstream Ascenda API unavailable"
          }
        }
      }
    },
    "/api/hotels/{id}/price": {
      "get": {
        "summary": "Fetch price for a specific hotel",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "MU74"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "destination_id",
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "EzoR"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "checkin",
            "schema": {
              "type": "string",
              "pattern": "^(0[1-9]|1[0-9]|20)[0-9]{2}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$",
              "example": "2026-08-24",
              "description": "ISO date string"
            },
            "required": true,
            "description": "ISO date string"
          },
          {
            "in": "query",
            "name": "checkout",
            "schema": {
              "type": "string",
              "pattern": "^(0[1-9]|1[0-9]|20)[0-9]{2}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$",
              "example": "2026-08-26",
              "description": "ISO date string"
            },
            "required": true,
            "description": "ISO date string"
          },
          {
            "in": "query",
            "name": "guests",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "2",
              "description": "Total guest count"
            },
            "required": true,
            "description": "Total guest count"
          },
          {
            "in": "query",
            "name": "currency",
            "schema": {
              "default": "SGD",
              "example": "SGD",
              "type": "string",
              "pattern": "^[A-Z]{3}$"
            }
          },
          {
            "in": "query",
            "name": "country_code",
            "schema": {
              "default": "SG",
              "example": "SG",
              "type": "string",
              "pattern": "^[A-Z]{2}$"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "default": "en_US",
              "example": "en_US",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel price data"
          },
          "400": {
            "description": "Invalid parameters or dates"
          },
          "502": {
            "description": "Upstream Ascenda API unavailable"
          }
        }
      }
    },
    "/api/webhooks/stripe": {
      "post": {
        "summary": "Stripe webhook endpoint",
        "tags": [
          "Webhooks"
        ],
        "description": "Receives Stripe payment events. Verified via stripe-signature header.",
        "parameters": [
          {
            "in": "header",
            "name": "stripe-signature",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Stripe webhook signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object",
                    "properties": {
                      "object": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id"
                        ]
                      }
                    },
                    "required": [
                      "object"
                    ]
                  }
                },
                "required": [
                  "type",
                  "data"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event received"
          },
          "400": {
            "description": "Missing or invalid stripe-signature header"
          },
          "401": {
            "description": "Invalid or missing Stripe signature"
          }
        }
      }
    }
  }
}